dws: factor out storage resource code into module
Problem: `coral2_dws.py` is too long and complicated and should have
its functionality broken up.

Break out functionality related to monitoring `Storage` resources
into a separate module.
